raw_requirements.txt
uvicorn
fastapi
cryptography
python-jose
orjson
ecdsa
tensorflow
torch
mmh3
pydantic_settings
sqlalchemy
python-multipart
pillow
openpyxl
colorlog
pytest
httpx